% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{data1}
\alias{data1}
\alias{data2}
\alias{data3}
\title{Some random data}
\format{A numeric matrix}
\source{
Generated from \code{set.seed(1); matrix(rnorm(10), ncol=2)}
}
\description{
A few numbers from \code{rnorm()} put into a matrix.
}
\examples{
data(data1)
str(data1)
}

