% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hiNorm}
\alias{hiNorm}
\title{Some random data}
\format{
A numeric vector.
}
\source{
Generated from \code{set.seed(123); rnorm(10)}
}
\description{
A few numbers from \code{rnorm()}.
}
\examples{
data(hiNorm)
str(hiNorm)
}
