% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C.R
\name{reverse_int}
\alias{reverse_int}
\title{Reverse an integer vector}
\usage{
reverse_int(x)
}
\arguments{
\item{x}{An integer vector}
}
\value{
An integer vector (reversed \code{x})
}
\description{
This function takes an integer vector and reverses it (in C).
}
\author{
Michael Chirico
}
