% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{rock_fun}
\alias{rock_fun}
\title{Scrolling letters}
\usage{
rock_fun(n = 20)
}
\arguments{
\item{n}{length of letters}
}
\value{
The last vector of letters.
}
\description{
Nothing fancy.
}
\examples{
library(rmini)
rock_fun()
rock_fun(50)
# you can call unexported functions with ::: but it is not recommended because
# package authors may not want you to go into their backyards
rmini:::add_one(c("a", "b", "c"))
}
