\name{split_filename}
\alias{split_filename}
\title{First line is title}
\usage{
split_filename(x, ...)
}
\arguments{
  \item{x}{explanation of \code{x}}

  \item{...}{explanation of \code{...}}
}
\value{
The value returned by this function.
}
\description{
This is a brief description.
}
\details{
Anything else after the description goes to the Details section.

You can write several paragraphs.
}
\examples{
split_filename("foo.bar")
}
\author{
Who are you?
}
\references{
\url{https://github.com/yihui/rmini}
}
\seealso{
\code{\link[tools]{file_ext}}, \code{\link[tools]{file_path_sans_ext}}
}
